requirements.txt
# Medical AI Project - Dependencies
# For Google Colab Pro+ GPU Training

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Medical Imaging
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Interpretability
pytorch-grad-cam>=1.4.8

# Chest X-Ray Classification
# (Uses ResNet50/EfficientNet with pretrained weights)

# Skin Lesion Detection
efficientnet-pytorch>=0.7.1

# Drug Discovery (Graph Neural Networks)
torch-geometric>=2.3.0
rdkit-pypi>=2022.9.5
deepchem>=2.7.0

# Utilities
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Advanced features
# grad-cam>=1.4.8
# captum>=0.6.0  # For advanced interpretability
# optuna>=3.3.0  # For hyperparameter optimization
